+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Updated
+- debril/feed-io to version 6
+
 ## [0.2.0] - 2022-06-10
 
 ### Added